requirements.txt
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.0
xlrd>=2.0.1
streamlit>=1.43.0
plotly>=5.24.0
fpdf2>=2.8.0
pylint>=3.3.0
pyright>=1.1.390
